+++ typescript/agentkit/src/wallet-providers/index.ts
@@ -7,3 +7,4 @@ export * from "./solanaKeypairWalletProvider";
 export * from "./privyWalletProvider";
 export * from "./privyEvmWalletProvider";
 export * from "./privySvmWalletProvider";
+export { PrivyWalletExport } from "./privyShared";
