+++ typescript/examples/langchain-privy-chatbot/chatbot.ts
@@ -7,10 +7,10 @@ import {
   pythActionProvider,
   cdpApiActionProvider,
 } from "@coinbase/agentkit";
-import { startAgent } from "./src";
-import { initializeWalletProvider } from "./src/agent";
+import { startAgent } from "./src/agent";
+import { initializeWalletProvider } from "./src/config";
+import { validateEnvironment } from "./src/env";
 
-const MODEL = "gpt-4o-mini";
 const MESSAGE_MODIFIER = `
   You are a helpful agent with a Privy server wallet that can interact onchain using the Coinbase Developer
   Platform AgentKit. You are empowered to interact onchain using your tools. If you ever need funds, you can
@@ -22,16 +22,22 @@ const MESSAGE_MODIFIER = `
   docs.cdp.coinbase.com for more information. Be concise and helpful with your responses. Refrain from
   restating your tools' descriptions unless it is explicitly requested.
 `;
-const THREAD_ID = "Privy AgentKit Chatbot Example!";
+const MODEL = "gpt-4o-mini";
+const NAME = "Privy AgentKit Chatbot Example!";
+const WALLET_DATA_FILE_PATH = "wallet_data.txt";
+const WALLET_DEFAULT_CHAIN_ID = "84532"; // base-sepolia
+const REQUIRED_ENV = [
+  "CDP_API_KEY_NAME",
+  "CDP_API_KEY_PRIVATE_KEY",
+  "OPENAI_API_KEY",
+  "PRIVY_APP_ID",
+  "PRIVY_APP_SECRET",
+] as const;
 
 /**
  * Get the default action providers for the agent
  */
 function getActionProviders() {
-  if (!process.env.CDP_API_KEY_NAME || !process.env.CDP_API_KEY_PRIVATE_KEY) {
-    throw new Error("CDP API key environment variables are not set");
-  }
-
   return [
     wethActionProvider(),
     pythActionProvider(),
@@ -50,7 +56,10 @@ function getActionProviders() {
  */
 export async function startPrivyChatbot(): Promise<void> {
   try {
-    const walletProvider = await initializeWalletProvider();
+    const walletProvider = await initializeWalletProvider({
+      walletDataFilePath: WALLET_DATA_FILE_PATH,
+      walletDefaultChainId: WALLET_DEFAULT_CHAIN_ID,
+    });
     const actionProviders = getActionProviders();
 
     await startAgent(
@@ -59,7 +68,7 @@ export async function startPrivyChatbot(): Promise<void> {
       {
         model: MODEL,
         messageModifier: MESSAGE_MODIFIER,
-        threadId: THREAD_ID,
+        threadId: NAME,
       }
     );
   } catch (error) {
@@ -72,6 +81,7 @@ export async function startPrivyChatbot(): Promise<void> {
 
 if (require.main === module) {
   console.log("Starting Privy Chatbot...");
+  validateEnvironment(REQUIRED_ENV);
   startPrivyChatbot().catch(error => {
     console.error("Fatal error:", error);
     process.exit(1);
