+++ typescript/examples/langchain-privy-chatbot/package.json
@@ -5,12 +5,13 @@
   "author": "Coinbase Inc.",
   "license": "Apache-2.0",
   "scripts": {
-    "start": "NODE_OPTIONS='--no-warnings' ts-node ./chatbot.ts",
+    "start": "NODE_OPTIONS='--no-warnings' ts-node chatbot.ts",
     "dev": "nodemon ./chatbot.ts",
-    "lint": "eslint -c .eslintrc.json *.ts",
-    "lint-fix": "eslint -c .eslintrc.json *.ts --fix",
     "format": "prettier -c .prettierrc --write \"**/*.{ts,js,cjs,json,md}\"",
-    "format-check": "prettier -c .prettierrc --check \"**/*.{ts,js,cjs,json,md}\""
+    "format:check": "prettier -c .prettierrc --check \"**/*.{ts,js,cjs,json,md}\"",
+    "lint": "eslint -c .eslintrc.json \"**/*.ts\"",
+    "lint:fix": "eslint -c .eslintrc.json \"**/*.ts\" --fix",
+    "test": "npx jest --no-cache --testMatch='**/*.test.ts'"
   },
   "dependencies": {
     "@coinbase/agentkit": "^0.2.2",
